Extract import patcher factory to test helpers
